#  remove caching for particle indices in cut region. fixes #2104 