In a chunked read (e.g. what happens when finding the value of a derived quantity), the size of a data object can change depending on which chunk we're looking at. This breaks the caching in the `_part_ind` property of the cut region data object. My quick and dirty fix is just to get rid of the caching. I *could* cache based on the underlying data chunk, which will be unique for each chunk, but then I'd blow up the memory requirements for doing any read on a cut_region, since that will require storing a mask for each chunk.

Suggestions from e.g. @brittonsmith would be great, particularly if the caching is a big performance win for the clump finder. Perhaps we could do the caching over there instead of on the cut region data object itself?