@brittonsmith 2018-11-30
@ngoldbaum, I have a reasonably large test case I can run. I'll do that and report back.