@brittonsmith 2018-12-05
I've done a reasonably large clump-finding run and here are some timing results:

Before this PR:
- clump-finding: 8265 seconds
- write-out: 5 seconds

After this PR:
- 8426 seconds
- write-out: 20 seconds

The time difference in the clump finding is ~2% and probably close to the run-to-run variation. The write-out time is simply the time to run `master_clump.save_as_dataset()` with particle fields. The factor of 4 increase is probably real, but it's a negligible fraction of the total time either way, so I can live with it. The test I ran was a pretty large example, so I think most use cases wouldn't see a big impact.